send things to Adri
