+++ assets/js/api.js
@@ -3,9 +3,9 @@
 function getPost() {
 return fetch('http://localhost:3000/posts')
 .then(response => response.json())
-.then(data => data)
+.then(data => console.log(data))
 }
 
+getPost();
 
-
-export {getPost}
+export { getPost };
